# docs/swagger.yaml
basePath: /v1
definitions:
...
  request.UserInput:
    properties:
      login:
        example: john_doe
        type: string
      password:
        example: s3cureP@ss
        type: string
    type: object
...
host: localhost:8080
info:
  contact: {}
  description: |-
    GophKeeper — secure vault for passwords, text notes, binary data and bank cards.

    ## Security
    All connections are encrypted with mandatory TLS.
    JWT tokens and all data are always protected in transit.
    Data at rest is encrypted per-user with AES-256-GCM.
    Encryption keys are derived from user passwords using Argon2id.
  title: Gophkeeper API
  version: "1.0"
paths:
  /api/user/delete-user:
    delete:
      consumes:
      - application/json
      description: Deletes the currently authenticated user's account
      produces:
      - application/json
      responses:
        "200":
          description: message
          schema:
            additionalProperties:
              type: string
            type: object
        "401":
          description: unauthorized
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: internal server error
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Delete user
      tags:
      - auth
  /api/user/login:
    post:
      consumes:
      - application/json
      description: Authenticates a user by login and password, returns a signed JWT
        token
      parameters:
      - description: User credentials
        in: body
        name: input
        required: true
        schema:
          $ref: '#/definitions/request.UserInput'
      produces:
      - application/json
      responses:
        "200":
          description: token
          schema:
            additionalProperties:
              type: string
            type: object
        "401":
          description: unauthorized
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: internal server error
          schema:
            additionalProperties:
              type: string
            type: object
      summary: User login
      tags:
      - auth
  /api/user/register:
    post:
      consumes:
      - application/json
      description: Creates a new user account, authenticates it and returns a JWT
        token
      parameters:
      - description: New user credentials
        in: body
        name: input
        required: true
        schema:
          $ref: '#/definitions/request.UserInput'
      produces:
      - application/json
      responses:
        "200":
          description: message
          schema:
            additionalProperties:
              type: string
            type: object
        "409":
          description: user already exists
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: internal server error
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Register user
      tags:
      - auth
...
securityDefinitions:
  BearerAuth:
    description: 'JWT token obtained from /api/user/login. Format: "Bearer {token}"'
    in: header
    name: Authorization
    type: apiKey
schemes:
- https
swagger: "2.0"